+++ include/simplebluez/interfaces/GattCharacteristic1.h
@@ -20,7 +20,7 @@ class GattCharacteristic1 : public SimpleDBus::Interface {
     void StartNotify();
     void StopNotify();
     void WriteValue(const ByteStrArray& value, WriteType type);
-    void WriteValue(const ByteArray value, const int size, WriteType type);
+    void WriteValue(const ByteArray value, const size_t size, WriteType type);
     ByteStrArray ReadValue();
 
     // ----- PROPERTIES -----
